Speed up `org-brain-headline-entries`.
When there are very many notes with IDs, it is a LOT faster to iterate through
all of the org entries in all org-brain-files than it is to rifle through
`org-id-locations` and find the entries it points to one-by-one.

On my computer, this patch reduces execution time from 10s to 1.7s.