Add `org-brain-headline-links-only-show-visible` customization variable
When this option is non-nil, links in entry headlines are replaced with the part
of the link that would be visible in an org-mode buffer. So, links will either
be displayed as the link's description (if there is one), or the link
destination.

This option is kept `nil` currently, so default behavior is not changed. I
think it would be good to enable it by default if it works well for others.