Add `org-brain-headline-links-only-show-visible` customization variable (#84)
* Add `org-brain-headline-links-only-show-visible` customization variable

When this option is non-nil, links in entry headlines are replaced with the part
of the link that would be visible in an org-mode buffer. So, links will either
be displayed as the link's description (if there is one), or the link
destination.

This option is kept `nil` currently, so default behavior is not changed. I
think it would be good to enable it by default if it works well for others.

* FIX: consistently use `pom` argument to `org-brain-headline-at`

* Make `pom` argument to `org-brain-headline-at` optional, as I meant to originally

* FIX bad indentation (spaces, not tabs)

* FIX STYLE: wrong symbol in docstring, & args in docstring should be upper-case

* FIX: Only call `org-brain-replace-links-with-visible-parts`, conditionally

That is done in `org-brain-headline-at` already, which respects the
`org-brain-headline-links-only-show-visible` customization var.