Create capture templates for org-brain
I plan to use these templates to implement functionality
`org-brain-add-todo` in a future commit. The summary of changes in
this commit are as follows:

1. Set up capture templates for brain.

I'm using "b" as a template prefix key for capture templates related
to org-brain. Since people may already be using this key in their own
capture templates, this functionality is turned off by default (via
`org-brain-visualize-use-capture-templates` var) and the user can
select a different prefix key (via
`org-brain-visualize-capture-prefix-key` var)

Since I was setting up capture templates, I also moved the capture
template described in the README to this code.

Finally, these templates only work in the context of
org-brain-visualize, to avoid clutter in other modes.

2. `org-brain-visualize--register-capture-templates`

This is a helper function to ensure that we set up the capture
templates only once. It's internal and invisible to the user.