Add tasks to the org-brain entry in visualize mode
Add a key-binding (`C-c t`) to directly add a TODO entry to the
current entry in org-brain. This makes it possible to capture action
items about the current entry without leaving the
org-brain-visualization window.

This change depends on the capture templates introduced in the commit
before it and adds `org-brain-visualize-add-todo` and
`org-brain-add-todo` functions to enable adding tasks under a brain
entry.