Add a note to an org-brain entry
Add a key-binding (`C-c z`) to directly trigger adding notes to the
current entry in org-brain. This makes it possible to capture thoughts
about the current entry without leaving the org-brain-visualization
window.

The commit provides two functions:

- `org-brain-visualize-add-note` : This function is bound to `C-c z`
in `org-brain-visualize-mode-map` and is the expected way to use this
functionality. Adds a note to the current entry.

- `org-brain-add-note`: This is a generic function that can be called
  from within org-mode.