Fix: Remove unnecessary * in plain capture template
The * causes a new heading to be created, under the heading that we
are already creating / accessing. The intention of the capture
template is to add content to an existing entry, or create a new entry
and add content to it. Both of these are done by the fixed template
now.